+++ CHANGELOG.md
@@ -3,6 +3,7 @@ CHANGELOG for Sulu
 
 * dev-develop
 
+    * BUGFIX      #571 [CoreBundle]     Fixed build command
     * ENHANCEMENT #556 [MediaBundle]    Enhanced url generation for collections
     * ENHANCEMENT #535 [SecurityBundle] Added SecurityListener and secured Content, Media and Security
     * BUGFIX      #568 [SnippetBundle]  Added template to view for Snippets